denoising-autoencoder-image-classifier/requirements.txt
tensorflow
numpy
pillow
gradio